Feature/24 : 경매 입찰하기

+++ src/main/kotlin/com/get_offer/product/service/ProductEditDto.kt
@@ -2,6 +2,7 @@ package com.get_offer.product.service
 
 import com.get_offer.product.controller.ProductEditReqDto
 import com.get_offer.product.domain.Category
+import java.math.BigDecimal
 import java.time.LocalDateTime
 import org.springframework.web.multipart.MultipartFile
 
@@ -11,7 +12,7 @@ data class ProductEditDto(
     val title: String? = null,
     val category: Category? = null,
     val description: String? = null,
-    val startPrice: Int? = null,
+    val startPrice: BigDecimal? = null,
     val startDate: LocalDateTime? = null,
     val endDate: LocalDateTime? = null,
     val images: List<MultipartFile>? = null
