+++ src/main/kotlin/com/get_offer/product/service/ProductListDto.kt
@@ -3,6 +3,7 @@ package com.get_offer.product.service
 import com.get_offer.product.domain.Category
 import com.get_offer.product.domain.Product
 import com.get_offer.product.domain.ProductStatus
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class ProductListDto(
@@ -11,7 +12,7 @@ data class ProductListDto(
     val name: String,
     val category: Category,
     val thumbnail: String,
-    val currentPrice: Int,
+    val currentPrice: BigDecimal,
     val status: ProductStatus,
     val startDate: LocalDateTime,
     val endDate: LocalDateTime,
